+++ src/dune/package.mli
@@ -45,6 +45,7 @@ module Dependency : sig
     type t =
       | Bvar of Var.t
       | Uop of Op.t * Var.t
+      | Bop of Op.t * Var.t * Var.t
       | And of t list
       | Or of t list
   end